can navigate to profile page and logout + small fixes
