+++ src/Pages/Login.jsx
@@ -1,0 +1,53 @@
+import { useState } from "react";
+import { useNavigate } from "react-router-dom";
+import { useAuth } from "../context/AuthContext";
+import Navbar from "../Components/Navbar";
+import "./Login.css";
+
+export default function Login() {
+  const [username, setUsername] = useState("");
+  const [password, setPassword] = useState("");
+  const [error, setError] = useState("");
+  const navigate = useNavigate();
+  const { login } = useAuth();
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    setError("");
+
+    if (login(username, password)) {
+      navigate("/");
+    } else {
+      setError("Please enter both username and password");
+    }
+  };
+
+  return (
+    <div className="page">
+      <Navbar />
+      <div className="login-container">
+        <div className="login-box">
+          <h1>Arcade Hub</h1>
+          <form onSubmit={handleSubmit}>
+            <input
+              type="text"
+              placeholder="Username"
+              value={username}
+              onChange={(e) => setUsername(e.target.value)}
+              required
+            />
+            <input
+              type="password"
+              placeholder="Password"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              required
+            />
+            {error && <p className="error-message">{error}</p>}
+            <button type="submit">Enter</button>
+          </form>
+        </div>
+      </div>
+    </div>
+  );
+}
