+++ src/pacman/pacman.css
@@ -1,3 +1,10 @@
+.page {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+}
+
 .game-container {
   display: flex;
   flex-direction: column;
